Refactored forwarding decorator, and extracted to it's own python file. Bug: Get container info is not array and will fail on adding IP on sub fields of array, when merging results from multiple instances
